% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hottest100-package.R
\docType{package}
\name{hottest100-package}
\alias{hottest100}
\alias{hottest100-package}
\title{hottest100: Compendium of Triple J's Hottest 100 tracks}
\description{
This is a data package containing records from \url{https://www.abc.net.au/triplej/countdown/hottest100/australiansongs} in November 2025.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katerobsau/hottest100}
}

}
\author{
\strong{Maintainer}: Kate Saunders \email{kate.saunders@monash.edu} (\href{https://orcid.org/0000-0002-1436-7802}{ORCID})

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})
  \item Cynthia A. Huang \email{cynthiahqy@gmail.com} (\href{https://orcid.org/0000-0002-9218-987X}{ORCID})
  \item Dianne Cook \email{visnut@gmail.com} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})
  \item Margaret Colville
}

}
\keyword{internal}
