% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polls}
\alias{polls}
\title{Triple J Hottest 100 Polls}
\format{
A data.frame with one row per countdown, including variables such as:
\describe{
  \item{year}{The year the poll was held.}
  \item{alltime}{Logical; whether the poll was an "All-Time" countdown}
  \item{alltimedescription}{A description of the poll, if applicable.}
  \item{colorRGB}{Primary RGB color value associated with the year in character format.}
  \item{contrastColorRGB}{Contrasting RGB color value (in character format) suitable for text overlay.}
}
}
\source{
\url{https://www.abc.net.au/triplej/hottest100/archive/search/}
}
\usage{
polls
}
\description{
A dataset with information about each Triple J Hottest 100 poll. Scraped on 22 Nov, 2025 from the official Triple J archive.
}
\keyword{datasets}
