% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{Information about tracks on the TripleJ Hottest 100}
\format{
A 4200 x 8 data frame
}
\description{
\itemize{
  \item alltime
.  \item artist
  \item country
  \item id
  \item pollyear
.  \item position
  \item releaseyear
  \item track
}
}
\keyword{datasets}
