% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albums}
\alias{albums}
\title{Information about albums on the TripleJ Hottest 100}
\format{
A 42 x 5 data frame
}
\description{
\itemize{
  \item year
  \item alltime
  \item alltimedescription
  \item colorRGB
  \item contrastColorRGB
}
}
\keyword{datasets}
