% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artists}
\alias{artists}
\title{Information about artists on the TripleJ Hottest 100}
\format{
A 2729 x 3 data frame
}
\description{
\itemize{
  \item artist
  \item country
  \item releaseyear
}
}
\keyword{datasets}
