% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artists}
\alias{artists}
\title{Triple J Hottest 100 Artists}
\format{
A data.frame with one row per artist, including variables:
\describe{
  \item{artist}{Artist name}
  \item{country}{Country of origin}
  \item{releaseyear}{Release year of included track}
}
}
\source{
Derived from \code{tracks}
}
\usage{
artists
}
\description{
A deduplicated dataset of artists represented in the Triple J Hottest 100,
  created from \code{tracks}.
}
\examples{
library(dplyr)
as_tibble(artists)
# who has appeared 10 or more times?
artists |>
  count(artist, sort = TRUE) |>
  filter(n >= 10)
# top 10 artists?
artists |>
  count(country, sort = TRUE) |>
  slice_max(n = 10, order_by = n)
}
\keyword{datasets}
