% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_is_the_hottest.R
\name{who_is_the_hottest}
\alias{who_is_the_hottest}
\title{Find the best in any year}
\usage{
who_is_the_hottest(year, special = NULL)
}
\arguments{
\item{year}{integer, pollyear to search in `tracks` data.}

\item{special}{logical, whether to include `alltime` in the data. Default is
to not include this.}
}
\description{
This function will help searching for information by year
}
\examples{
who_is_the_hottest(2023)
}
